+++ src/index.ts
@@ -7,7 +7,7 @@
  */
 
 export { delegate, delegateCapture } from './delegate.js';
-export { raw, SafeHtml } from './jsx-runtime.js';
+export { isSafeHtml, raw, SafeHtml } from './jsx-runtime.js';
 export { mount } from './mount.js';
 export {
   batch,